Add RequestMixin to define common make_request method
