+++ seam/__init__.py
@@ -1,7 +1,9 @@
 # flake8: noqa
 # type: ignore
 
-from seam.seam import Seam, SeamApiException
+from seam.seam import Seam
+from seam.types import SeamApiException
+from seam.seam_multi_workspace import SeamMultiWorkspace
 from seam.options import SeamHttpInvalidOptionsError
 from seam.auth import SeamHttpInvalidTokenError
 from seam.utils.action_attempt_errors import (
